requirements.txt
amqp==2.5.2
appnope==0.1.0
asgiref==3.2.3
attrs==19.3.0
Automat==0.8.0
backcall==0.1.0
beautifulsoup4==4.8.2
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
decorator==4.4.1
Django==3.0.2
django-bootstrap4==1.1.1
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.3.0
incremental==17.5.0
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
kombu==4.6.7
lxml==4.4.2
more-itertools==8.0.2
parsel==1.5.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
pipenv==2018.11.26
prompt-toolkit==3.0.2
Protego==0.1.16
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.5.2
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
PySocks==1.7.1
python-anticaptcha==0.4.2
pytz==2019.3
queuelib==1.5.0
redis==3.3.11
requests==2.22.0
Scrapy==1.8.0
scrapy-djangoitem==1.1.1
scrapy-inline-requests==0.3.1
service-identity==18.1.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
stem==1.8.0
traitlets==4.3.3
Twisted==19.10.0
urllib3==1.25.7
vine==1.3.0
virtualenv==16.7.9
virtualenv-clone==0.5.3
w3lib==1.21.0
wcwidth==0.1.8
zipp==0.6.0
zope.interface==4.7.1
